refactor : PostService의 getPopularPost 메서드를 GetPopularPostFacade 클래스로 이동
